json: include header data in parsed body. (#231)
As it is now, the generated json services only gets the returned body
data when constructing the output object/shape. This will always give
members located in response headers a null value. In a future breaking
change, the return type can be changed to Map<String, dynamic>, but for
now, this will suffice.